backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.0
xgboost==2.0.0
tensorflow==2.13.0
pandas==2.1.0
numpy==1.24.0
joblib==1.3.0
Pillow==10.0.0
httpx==0.25.0
pymongo==4.6.0
python-dotenv==1.0.0
anthropic==0.7.0
python-multipart==0.0.6
